refac: 유저할당 job 역할 클래스로 분리
